chore: enable strict compiler flags for tests and integration (#65)
* chore: enable -Wall -Wextra -Werror for GCC/Clang and /W4 /WX for MSVC


* fix: resolve compiler warnings surfaced by -Wall -Wextra

- Aeu.h: fix signed/unsigned comparisons in shift and compareTo operators
- Aeu.h: remove unused tX/tY variables from gcd loop
- Aeu.h: replace characters[] array with Char() literals to fix unused lambda capture
- Aesi.h: add parentheses to requires clause (&&-within-||)
- Aesi.h: mark traverseDashes specializations inline to avoid unused-function warning
- display.cpp: remove unused size return value captures from getString calls


* ci: run coverage pipeline in Release mode


* fix: resolve MSVC C4244 narrowing conversion warnings in Aeu.h

Add explicit static_cast<block> and static_cast<Unsigned> for
uint64_t → uint32_t/Unsigned assignments in operator+= and operator*=
to silence C4244 errors under /W4 /WX on MSVC.


* fix: rename local variable 'block' to avoid shadowing the 'block' type

In operator*= (uint64_t and Aeu overloads), the local variable named
'block' shadowed the type alias 'block', causing compilation errors on
GCC, Clang, and MSVC.


* chore: enable -Wconversion -Wsign-conversion for GCC/Clang


* fix: resolve -Wconversion -Wsign-conversion warnings

Replace long long loop counters with std::size_t using the i-- > 0
idiom for safe reverse iteration. Fix signed/unsigned arithmetic in
Aesi integral constructors and operators. Use 1ULL for bit-shift stamp
calculations to avoid UB when shift amount equals type width.


* fix: resolve GCC sign-conversion and narrowing warnings

Use unsigned long long for Aeu constructor's accumulator to avoid
implicit uint64_t/unsigned long long mismatch on Linux (where
uint64_t = unsigned long). Cast char arithmetic result back to Char
in getString to silence int-to-char narrowing warning on GCC.


* fix: resolve GCC sign-conversion warnings in Aesi compareTo integral overload

On Linux (LP64), uint64_t is 'unsigned long' while 'unsigned long long'
is a distinct type. Passing unsigned long long to compareTo(uint64_t)
caused an implicit conversion that GCC flagged as sign-conversion.
Cast through int64_t first (signed widening) then explicitly to uint64_t
to match the parameter type exactly and silence the warning.


* fix: resolve -Wconversion and -Wsign-conversion warnings in getString and wchar_t constructor


* chore: enable -Wpedantic for GCC/Clang


* chore: enable -Wshadow -Wold-style-cast for GCC/Clang


* fix: replace C-style casts with static_cast in test files


* chore: enable -Wnull-dereference -Wundef for GCC/Clang


* fix: remove extra semicolons after operator<=> definitions

GCC -Wpedantic treats trailing ';' after a function body inside a class
as an error.


* refactor: extract gmpToBinaryString helper to remove 6 duplicate blocks in display tests


* fix: use constant format string literals in snprintf/swprintf calls

Eliminates CWE-134 static analysis warnings by splitting the ternary
format string into explicit if/else branches with constant literals.


* refactor: make hexUppercase a template parameter in getString

Moves hexUppercase from a runtime bool parameter to a compile-time
template parameter, allowing if constexpr branches and eliminating
the CWE-134 taint path through the format string selection.


---------

+++ test/arithmetic/addition/addition.cpp
@@ -51,7 +51,7 @@ void testAdditionBasic() {
 template <typename T, std::size_t N>
 void runIncrementLoop(const mpz_class& l) {
     T value = l;
-    const std::size_t increments = rand() % 100;
+    const std::size_t increments = static_cast<std::size_t>(rand()) % 100;
     for (std::size_t j = 0; j < increments * 2; j += 2) {
         EXPECT_EQ(value++, l + j);
         EXPECT_EQ(++value, l + j + 2);
